requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
Flask==2.1.2
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.12.0
pyinstaller==5.0
pyinstaller-hooks-contrib==2022.4
pyparsing==3.0.8
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.28.1
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
Werkzeug==2.1.2
yarl==1.7.2
